requirements.txt
tsp-solver2==0.3
pandas==0.24.2
numpy==1.16.3